+++ src/DigiSign.php
@@ -9,6 +9,7 @@ use DigitalCz\DigiSign\Auth\CachedCredentials;
 use DigitalCz\DigiSign\Auth\Credentials;
 use DigitalCz\DigiSign\Endpoint\AccountEndpoint;
 use DigitalCz\DigiSign\Endpoint\AuthEndpoint;
+use DigitalCz\DigiSign\Endpoint\BatchSendingsEndpoint;
 use DigitalCz\DigiSign\Endpoint\DeliveriesEndpoint;
 use DigitalCz\DigiSign\Endpoint\EndpointInterface;
 use DigitalCz\DigiSign\Endpoint\EnumsEndpoint;
@@ -128,8 +129,8 @@ final class DigiSign implements EndpointInterface
             throw new InvalidSignatureException('Unable to parse signature header');
         }
 
-        $ts = (int)($matches['t'] ?? 0);
-        $signature = $matches['s'] ?? '';
+        $ts = (int)($matches['t']);
+        $signature = $matches['s'];
 
         if ($ts < time() - $this->signatureTolerance) {
             throw new InvalidSignatureException("Request is older than {$this->signatureTolerance} seconds");
@@ -241,6 +242,11 @@ final class DigiSign implements EndpointInterface
     public function deliveries(): DeliveriesEndpoint
     {
         return new DeliveriesEndpoint($this);
+    }
+
+    public function batchSendings(): BatchSendingsEndpoint
+    {
+        return new BatchSendingsEndpoint($this);
     }
 
     public function identifications(): IdentificationsEndpoint