+++ src/Resource/EnvelopeDocument.php
@@ -11,6 +11,8 @@ class EnvelopeDocument extends BaseResource
 {
     use EntityResourceTrait;
 
+    public string $id;
+
     public string $name;
 
     public ?string $metadata = null;