+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeDocument.assignments` to resource
 - Add `EnvelopeRecipient` signed document columns
 - Add `EnvelopeTagsByPlaceholder` endpoint
+- Add `EnvelopeRecipient.certificateInfo` endpoint
 
 ## [2.3.0] - 2024-03-28
 ### Added