+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 - Add `MyIdentifications.info` endpoint
+- Add `EnvelopeRecipient.certificateInfo` endpoint
 
 ## [2.4.0] - 2024-05-14
 ### Added
@@ -24,7 +25,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeDocument.assignments` to resource
 - Add `EnvelopeRecipient` signed document columns
 - Add `EnvelopeTagsByPlaceholder` endpoint
-- Add `EnvelopeRecipient.certificateInfo` endpoint
 
 ## [2.3.0] - 2024-03-28
 ### Added