+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `MyIdentifications.info` endpoint
 - Add `EnvelopeRecipient.certificateInfo` endpoint
 - Add `EnvelopeProperties.channelForSender` in to resource
+- Edit `CertificateInfo` resource
 
 ## [2.4.0] - 2024-05-14
 ### Added