+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `EnvelopeRecipient.certificateInfo` endpoint
 - Add `EnvelopeProperties.channelForSender` in to resource
 - Edit `CertificateInfo` resource
+- Extend `Contact` resource
 
 ## [2.4.0] - 2024-05-14
 ### Added