+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Extend `Contact` resource
 - Extend `AccountBilling` resource with identify limits
 - Add `AccountBilling.userContactLimit`
+- Add `IdentifyScenarioVersion.restrictedCountries`
 
 ## [2.4.0] - 2024-05-14
 ### Added