+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Add `AccountBilling.userContactLimit`
 - Add `IdentifyScenarioVersion.restrictedCountries`
 - Add `EnvelopeTemplateDocumentAssignments` endpoint
+- Add `Envelope.createdBy` and `Envelope.sentBy`
 
 ## [2.4.0] - 2024-05-14
 ### Added