+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [Unreleased]
+## [2.5.0] - 2024-10-14
 - Add `AccountBilling.automaticTagsPlacement`
 - Add `AccountBilling.batchSending`
 - Extend `EnvelopeRecipient` and `EnvelopeTemplateRecipient` resource